+++ tests/issues.rs
@@ -4,7 +4,7 @@
 
 use std::sync::mpsc;
 
-use quick_xml::errors::SyntaxError;
+use quick_xml::errors::{IllFormedError, SyntaxError};
 use quick_xml::events::{BytesDecl, BytesStart, BytesText, Event};
 use quick_xml::name::QName;
 use quick_xml::reader::Reader;
@@ -145,14 +145,14 @@ mod issue514 {
         reader.check_end_names(true);
 
         match reader.read_event() {
-            Err(Error::EndEventMismatch { expected, found }) => {
-                assert_eq!(expected, "some-tag");
-                assert_eq!(found, "other-tag");
-            }
-            x => panic!(
-                "Expected `Err(EndEventMismatch {{ expected = 'some-tag', found = 'other-tag' }})`, but got `{:?}`",
-                x
+            Err(Error::IllFormed(cause)) => assert_eq!(
+                cause,
+                IllFormedError::MismatchedEnd {
+                    expected: "some-tag".into(),
+                    found: "other-tag".into(),
+                }
             ),
+            x => panic!("Expected `Err(IllFormed(_))`, but got `{:?}`", x),
         }
         assert_eq!(reader.read_event().unwrap(), Event::Eof);
     }